Fix for 5914.
I am not quite happy with this fix though it does solve the problem. What we see is that if content is defined with valid JSON that JSON is encoded into a dict. The type error happens because the dict is not serialized and we are unable to write it.