inventory/__init__.py: Add localhost if in a subset(--limit)
Addresses GH-6449 and improves on GH-6479 which test against a playbook that addresses all using --limit=localhost or --limit='localhost:!127.0.0.1'.

We have this concept that localhost is implicitly added to the inventory if defined in a playbook. This commit extends that out to the subsets described by the limit option.